requirements.txt
# Веб-фреймворк и сервер
fastapi
uvicorn[standard]

# Работа с базой данных
SQLAlchemy

# Шаблонизатор для HTML
Jinja2

# Генерация PDF
reportlab

# Для работы с .env файлами (понадобится на следующем шаге)
python-dotenv

# Security and Authentication
passlib[bcrypt]

python-jose[cryptography]

pydantic[email]

python-multipart

PyMySQL

black

isort 

flake8